% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league_teams_meta.R
\name{get_league_teams_meta}
\alias{get_league_teams_meta}
\title{FUNCTION_TITLE}
\usage{
get_league_teams_meta(league, year)
}
\arguments{
\item{league}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  leagues_meta <- get_leagues_meta()
  leagues_meta
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{nest}}
}
