% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leagues_meta.R
\name{get_leagues_meta}
\alias{get_leagues_meta}
\title{FUNCTION_TITLE}
\usage{
get_leagues_meta()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  league_teams_meta <- get_league_teams_meta(league = 'fagprem', year = 2016L)
  league_teams_meta
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{arrange}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{nest}}
}
