% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_players.R
\name{get_team_players}
\alias{get_team_players}
\title{FUNCTION_TITLE}
\usage{
get_team_players(league, year, team)
}
\arguments{
\item{league}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{team}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
   # Works
   get_team_players(league = 'faprem', year = 2016L, team = 'Tottenham Hotspur')
   
   # Fails because 'engprem' only started to be used beginning in 2018
   get_team_players(league = 'engprem', year = 2016L, team = 'Tottenham Hotspur')
   
   # Fails (for reason mentioned above).
   get_team_players(league = 'faprem', year = 2018L, team = 'Tottenham Hotspur')
   
   # Fails due to invalid team name.
   get_team_players(league = 'faprem', year = 2016L, team = 'Tottenham')
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{nest}}
}
